qf: Move queries into own package (for namespacing)
